+++ modules/ModuleAnyStoresSearchMap.php
@@ -55,7 +55,10 @@ class ModuleAnyStoresSearchMap extends ModuleAnyStoresList
     {
         parent::compile();
 
-        $GLOBALS['TL_JAVASCRIPT']['googleapis-maps'] = 'https://maps.googleapis.com/maps/api/js?language=' . $GLOBALS['TL_LANGUAGE'];
+        // load Google Maps JavaScript
+        $arrParams = $this->anystores_signedIn ? array('signed_in' => 'true') : array();
+        GoogleMaps::includeJs($arrParams);
+
         $GLOBALS['TL_JAVASCRIPT']['markerclusterer'] = 'system/modules/anyStores/assets/js/markerclusterer/src/markerclusterer_compiled.js';
 
         // map height